backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
httpx==0.27.0

# LLM
openai==1.30.0
anthropic==0.28.0

# STT
deepgram-sdk==3.2.7

# TTS (ElevenLabs)
elevenlabs==1.1.2

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
pgvector==0.2.5

# Redis / session memory
redis[asyncio]==5.0.4

# Background jobs (outbound campaigns)
celery==5.3.6
celery[redis]==5.3.6

# Outbound calls
twilio==8.13.0

# Language detection
langdetect==1.0.9

# Monitoring
prometheus-client==0.20.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1
loguru==0.7.2
